requirements.txt
appdirs==1.4.4
asgiref==3.3.4
black==21.5b1
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.3
django-cors-headers==3.2.1
django-rest-swagger==2.2.0
djangorestframework==3.12.4
idna==2.10
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
mypy-extensions==0.4.3
openapi-codec==1.3.2
pathspec==0.8.1
psycopg2-binary==2.8.6
PyJWT==2.1.0
pytz==2021.1
regex==2021.4.4
requests==2.25.1
simplejson==3.17.2
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.4
whitenoise==5.2.0
